refactoring values that should never be fractional
https://trello.com/c/OFkODanH/55-refactor-value-types

changed all interactions from damage, health, and gold to be integers instead of floats. declarations and function returns.

divisions for progress bars cast the divisor into a float like so.
playerHealthSlider.value = health / (float)maxHealth;

probability is untouched

+++ PuzzleItOut/Assets/Scripts/CombatManager.cs
@@ -27,7 +27,7 @@ public class CombatManager : MonoBehaviour
 
     }
 
-    public void DealDamage(float damage)
+    public void DealDamage(int damage)
     {
         return;
     }
@@ -60,23 +60,23 @@ public class CombatManager : MonoBehaviour
         return -1;
     }
 
-    public float CalculateDamage(ComboScriptable combo, List<PieceScriptable> pieces)
+    public int CalculateDamage(ComboScriptable combo, List<PieceScriptable> pieces)
     {
-        float result = combo.Damage(pieces);
+        int result = combo.Damage(pieces);
         
         return result;
     }
 
-    public float CalculateGold(ComboScriptable combo, List<PieceScriptable> pieces)
+    public int CalculateGold(ComboScriptable combo, List<PieceScriptable> pieces)
     {
-        float result = combo.Gold(pieces);
+        int result = combo.Gold(pieces);
         //Debug.Log($"'{combo.comboName}' generated {result} gold");
         return result;
     }
 
-    public float CalculateHealth(ComboScriptable combo, List<PieceScriptable> pieces)
+    public int CalculateHealth(ComboScriptable combo, List<PieceScriptable> pieces)
     {
-        float result = combo.Health(pieces);
+        int result = combo.Health(pieces);
         //Debug.Log($"'{combo.comboName}' healed {result} health");
         return result;
     }
